Amends _bower.json to fix Sass punctuation problem
Previously, if you used the generator to make an app with Sass, the bower.json file was generated with an extra comma on line 5. This comma prevented bower install from working and had to be removed by hand. Now the comma is included in the if (less) block, so if you choose Sass the comma will not be inserted.